requirements.txt
torch
torchvision
matplotlib
numpy
opencv-python
pillow
wandb
tqdm
jupyter
ipykernel 